+++ lib/flutter_injections.dart
@@ -3,4 +3,6 @@ library flutter_injections;
 export 'src/flutter_injections.dart';
 export 'src/inject.dart';
 export 'src/module.dart';
+export 'src/navigator/flutter_app.dart';
+export 'src/navigator/flutter_navigator_module.dart';
 export 'src/widget.dart';